% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDPMIenrich.R
\name{TDPMIenrich}
\alias{TDPMIenrich}
\title{For time-dependent progression-related gene set enrichment analysis of myocardial infarction}
\usage{
TDPMIenrich(
  genes,
  pvalueFilter = 0.05,
  qvalueFilter = 0.05,
  highcol = "grey",
  lowcol = "red3",
  useFilter = "TRUE",
  typefigure = "bubble",
  organism = "hsa"
)
}
\arguments{
\item{genes}{a vector of a list of genes}

\item{pvalueFilter}{pvalue filter for enrichment}

\item{qvalueFilter}{qvalue filter for enrichment}

\item{highcol}{color for high statistical significance}

\item{lowcol}{color for low statistical significance}

\item{useFilter}{Whether to filter the results based on the pvalue and qvalue}

\item{typefigure}{The type of the output figure, including "barplot", "bubble" and "all"}

\item{organism}{the organism of input genes (using "hsa" or "mmu")}
}
\value{
a dataframe of TDPMI enrichment result
}
\description{
For time-dependent progression-related gene set enrichment analysis of myocardial infarction
}
\examples{
result=TDPMIenrich(genes=c("Aqp1","Cxcl3","Gm26870","Hba-a1","Hba-a2","Hbb-bs","Hbb-bt","Mb","mt-Atp6","mt-Atp8","mt-Co1","mt-Co2"),organism = "mmu",pvalueFilter=0.05,qvalueFilter=0.05,highcol="grey",lowcol="red3",useFilter="TRUE",typefigure="bubble")
}
